LF-14261-8 net: enetc: set the command cache attribute for i.MX ENETC SI
commit  0a839675ea35b2288a098d2f030438d6bbed429d from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git linux_6.12.3-imx943-er1

We have set SI's buffer cache attribute for Tx/Rx BD ring, but forgot to
set the cache attribute for the command BD ring. The command BD ring is
similar to Tx/Rx BD ring, on the LS1028A platform, the value of command
cache attribute is the same as that of buffer cache attribute. To keep
it consistent, it is also set to the same value on the i.MX platforms.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>