LF-13889-2: media: amphion: Add a frame flush mode for decoder
commit  170abf74ca87d538d72d864498b5f729481f0d77 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The amphion decoder will pre-parse 3 frames before decoding the first
frame. If we append a flush padding data after frame, the decoder
will finish parsing and start to decode when the flush data is parsed.
It can reduce the decoding latency.
In the past, we only enable this mode when the display delay is set to
0. But we still can enable this mode without changing the display order,
so we add a frame_flush_mode parameter to enable it.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: TaoJiang <tao.jiang_2@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>