LF-14123-01 soc: imx: imx93_lpm: Add cooling device support to lpm control
commit  3eaf09dd1279468f124fcf3a2cf82c9bfc4f33c2 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

lpm module has three states: suspend_prepared state, cooling_active
state, and manual_setting state, with their priority decreasing in
order.

When the system is about to go to sleep, lpm is in suspend_prepared
state, the mode is switched to OD/ND, depending on whether OD mode
is supported.

When the system is not about to sleep and the system temperature is
at the passive trip temperature, the lpm is in cooling_active state,
and the target mode of the cooling device is LD/ND, depending on
whether the LD mode is supported. However, if the manually set mode
would make the system cooler than the target mode of the cooling
device, the system remains in the manually set mode.

When the system is not in the above two states, lpm is in manual_setting
state, and the mode can be switched via the /sys interface.

Signed-off-by: Pengfei Li <pengfei.li_1@nxp.com>
Reviewed-by: Jacky Bai <ping.bai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>