LF-13188 spi: fspi: correct the clock rate limitation for different mode
commit  cc644f3b19fb328d3a907f8664001b1aedda0d23 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Currently, for DTR mode, we use mode 3 to select sample clock source,
according to the datasheet, for flexspi, the max frequency in this case
is 166MHz. For STR mode, we use mode 0 to select sample clock source,
according to the datasheet, for flexspi, the max frequency in this case
is 66MHz. So add the clock rate limitation accordingly.

Beside, for upper clock limitation, they are for the clock output to
device. For DTR mode, flexspi will automatically divide 2 from serial
root clock to get the output clock. So add the fix for this case when
config the serial root clock rate.

Reviewed-by: Han Xu <han.xu@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>