LF-14085 : trusted_dcp.c: Do not return in case of non-secure mode
commit  05aecc02527bd0fda3d62cf75bdfb6811e1f7fdb from
https://github.com/nxp-imx/linux-imx lf-6.12.y

There are multiple type of keys in different worlds, like
test key in case of non-secure world and OTP, unique key
in case of secure world.
So, instead of returning with an error, in case of test key, we
should display warning to the user and allow the user to run the
trusted key functionality with test key.

Signed-off-by: Kshitiz Varshney <kshitiz.varshney@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>