LF-13267: media: nxp: imx8-isi: check streaming status in streamon()
commit  59e3a3430258753e2ce7908abf1f286b4f811548 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Return 0 if streaming on the video device was already started. According
to the documentation:

"If VIDIOC_STREAMON is called when streaming is already in progress,
or if VIDIOC_STREAMOFF is called when streaming is already stopped,
then 0 is returned."

However, in our case -EBUSY is returned by mxc_isi_pipe_acquire(). Fix
this by checking whether we're streaming or not, before going through
with acquiring the ISI pipe.

Fixes: cf21f328fcaf ("media: nxp: Add i.MX8 ISI driver")
Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Guoniu Zhou <guoniu.zhou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>