LF-13782-3: arm64: dts: imx95-19x19-verdin: Fix suspend for os08a20 on csi-only
commit  4504a519cb4dc83cd8aed2ad5b6210854e6d9c4a from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The reset pin was not claimed because the name of the reset gpio
was wrong in the dts, rst-gpios versus reset-gpios. The sensor driver uses
devm_gpiod_get_optional which does not show error in this case.

Fixes: 48b51329e5a9 ("LF-12850-2: arm64: dts: imx95: os08a20: Update verdin dts to use the compatible from ox05b1s driver")

Signed-off-by: Mirela Rabulea <mirela.rabulea@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>