LF-13782-4: arm64: dts: imx95-19x19-verdin: Fix suspend for ox05b1s on csi-only
commit  7adf8c36ddfce7bf434f976d7ffb60be7b112d7a from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The reset pin was not claimed because the name of the reset gpio
was wrong in the dts, rst-gpios versus reset-gpios. The sensor driver uses
devm_gpiod_get_optional which does not show error in this case.

Fixes: 8965a0ed9690 ("LF-12885-1: arm64: dts: imx95-19x19: ox05b1s: Add ox05b1s dtb for verdin")

Signed-off-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>