LF-14228 clk: imx: Init clk flags if div or mux is present
commit  4f5419ad749db7e3a2ea8d9ff4df634fa1003424 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

For some of the PCC slots, the divider or mux control may
not present in HW. The 'CLK_SET_x_GATE' should only be added
when the corresponding control is present.

if the PCC slot only has gate control, no need to register
it with the 'CLK_SET_x_GATE' flags. with these flags, If the
clock is prepared, the clock itself and the parent will be
set into protected, if we want to change the parent clock
setting, the operation will be blocked.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>