LF-13948-5: media: imx8mq-mipi-csi2: fix get_fmt with streams
commit  f4d9334b98aae891fe62d5b8db0ec536a3337b9b from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Currently, the get_fmt does note care about the stream received on the
format, so it only set the format on stream 0. Fix this in order to be
able to set the format on all the exposed streams.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>