LF-13948-7: media: imx8mq-mipi-csi2: add remote_pad
commit  c2c67c4468afef00bcbf4462671a05dcf7142dc9 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Currently, the remote_pad is not stored correctly, causing
get_frame_desc to be called on the wrong remote pad.
This patch stores the remote pad on either link_setup (if the link is
manually created) or on notify_bound, when the link will be
automatically created.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>