LF-14261-10 net: enetc: allow untrusted VFs to disable promiscuous mode
commit  682939cabd49d896afd461e61fdaf4c17b8e773c from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git linux_6.12.3-imx943-er1

Untrusted VFs should not have permission to enable MAC promiscuous mode,
however, the hardware enables promiscuous mode for all SIs by default.
So the PF driver disables promiscuous mode for all SIs when it probes,
but it has no effect on VSI, because the VSI driver performs FLR when it
probes, which will cause the bits of PSIPMMR corresponding to VSI to be
reset. In other words, the promiscuous mode of the VSI will be enabled
due to FLR.

The most correct approach is for the PF driver to detect the FLR event
of the VSI and then take corresponding actions. However, the PF driver
does not currently support this event, so the simple approach is to
allow untrusted VFs to disable promiscuous mode, because promiscuous
mode is disabled by default when the VSI network device is up.

When the PF driver supports detecting VSI FLR event, it will disable
promiscuous mode for untrusted VFs in the event handling. Of course,
the current solution should also be retained, because untrusted VFs
just cannot enable promiscuous mode, and it is reasonable to disable
promiscuous mode.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>