LF-14260-29 net: dsa: netc: add initial PM support
commit  6972f34e92fbf61cee18b467ea3bb5858cd07061 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git linux_6.12.3-imx943-er1

Currently, suspend/resume of NETC Switch is not fully supported. This is
just a temporary patch to solve the call trace and crash problems caused
by Switch during system suspend/resume. There is no reason to block the
system suspend/resume process due to the switch problem. So this is a
quick fix and full support will be added in the future.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>