LF-14450 netc: enetc: do not dump BD rings when the port is down
commit  cd3f85649bda4ebbd897b7488e7fb8416f43c11b from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git linux_6.12.3-imx943-er1

The Tx and Rx BD rings are freed when the ENETC port is down, so if an
an attempt to dump the BD rings will lead to a NULL pointer access. To
avoid this, add netif_running() check before accessing the BD rings.

Fixes: 1ffd37d8a30b ("LF-10640-7 net: enetc: Add i.MX95 ENETC support")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>